requirements.txt
crewai
langchain_experimental
langchain_community
